# Add kineto and autograd (emit_nvtx) profiling