@jithunnair-amd 2022-06-06
Also added autograd emit_nvtx profiling using the `--autograd_profiler` flag (mutually exclusive and lower priority to `--kineto`). Ran with rpd_tracer as follows:
`runTracer.sh python micro_benchmarking_pytorch.py --network resnet50 --autograd_profiler`