@jithunnair-amd 2022-06-09
> Any confirmation whether the performance running without profiling was affected in any way by the PR changes?

@jeffdaily @mwootton Yes, I verified by running 1. without the `autograd.profiler.emit_nvtx` context manager and 2. with the context manager but with `enabled=False`, and I got the same performance for both cases.